Download results as .csv file
Added ability to download results as .csv file. Some of the code that enables this is from Stack Overflow under the CC BY-SA 3.0 license, which is now noted in LICENSE.txt, README.md, and the top comment in high_voltage_divider.html

Also fixed an input checking bug: The check on the minimum string resistance limited that value to the largest resistor in the chosen series. It is now three times as large.